+++ test/Microsoft.OpenApi.Readers.Tests/V2Tests/OpenApiSecuritySchemeTests.cs
@@ -23,7 +23,7 @@ namespace Microsoft.OpenApi.Readers.Tests.V2Tests
         {
             using (var stream = Resources.GetStream(Path.Combine(SampleFolderPath, "basicSecurityScheme.yaml")))
             {
-                var document = OpenApiStreamReader.LoadYamlDocument(stream);
+                var document = LoadYamlDocument(stream);
 
                 var context = new ParsingContext();
                 var diagnostic = new OpenApiDiagnostic();
@@ -48,7 +48,7 @@ namespace Microsoft.OpenApi.Readers.Tests.V2Tests
         {
             using (var stream = Resources.GetStream(Path.Combine(SampleFolderPath, "apiKeySecurityScheme.yaml")))
             {
-                var document = OpenApiStreamReader.LoadYamlDocument(stream);
+                var document = LoadYamlDocument(stream);
                 var context = new ParsingContext();
                 var diagnostic = new OpenApiDiagnostic();
 
@@ -73,7 +73,7 @@ namespace Microsoft.OpenApi.Readers.Tests.V2Tests
         {
             using (var stream = Resources.GetStream(Path.Combine(SampleFolderPath, "oauth2ImplicitSecurityScheme.yaml")))
             {
-                var document = OpenApiStreamReader.LoadYamlDocument(stream);
+                var document = LoadYamlDocument(stream);
                 var context = new ParsingContext();
                 var diagnostic = new OpenApiDiagnostic();
 
@@ -108,7 +108,7 @@ namespace Microsoft.OpenApi.Readers.Tests.V2Tests
         {
             using (var stream = Resources.GetStream(Path.Combine(SampleFolderPath, "oauth2PasswordSecurityScheme.yaml")))
             {
-                var document = OpenApiStreamReader.LoadYamlDocument(stream);
+                var document = LoadYamlDocument(stream);
                 var context = new ParsingContext();
                 var diagnostic = new OpenApiDiagnostic();
 
@@ -143,7 +143,7 @@ namespace Microsoft.OpenApi.Readers.Tests.V2Tests
         {
             using (var stream = Resources.GetStream(Path.Combine(SampleFolderPath, "oauth2ApplicationSecurityScheme.yaml")))
             {
-                var document = OpenApiStreamReader.LoadYamlDocument(stream);
+                var document = LoadYamlDocument(stream);
                 var context = new ParsingContext();
                 var diagnostic = new OpenApiDiagnostic();
 
@@ -178,7 +178,7 @@ namespace Microsoft.OpenApi.Readers.Tests.V2Tests
         {
             using (var stream = Resources.GetStream(Path.Combine(SampleFolderPath, "oauth2AccessCodeSecurityScheme.yaml")))
             {
-                var document = OpenApiStreamReader.LoadYamlDocument(stream);
+                var document = LoadYamlDocument(stream);
 
                 var context = new ParsingContext();
                 var diagnostic = new OpenApiDiagnostic();
@@ -206,6 +206,16 @@ namespace Microsoft.OpenApi.Readers.Tests.V2Tests
                             }
                         }
                     });
+            }
+        }
+
+        static YamlDocument LoadYamlDocument(Stream input)
+        {
+            using (var reader = new StreamReader(input))
+            {
+                var yamlStream = new YamlStream();
+                yamlStream.Load(reader);
+                return yamlStream.Documents.First();
             }
         }
     }
